+++ src/visualisations/SondeAnalogique.java
@@ -8,7 +8,7 @@
  * Classe réalisant l'affichage d'information composée d'éléments réels (float)
  * @author prou
  */
-   public class SondeAnalogique extends Sonde <Float> {
+   public class SondeAnalogique extends Sonde <Double> {
     
    
     /**
@@ -21,12 +21,12 @@
    
    
    	 
-      public void recevoir (Information <Float> information) { 
+      public void recevoir (Information <Double> information) { 
          informationRecue = information;
          int nbElements = information.nbElements();
-         float [] table = new float[nbElements];
+         double [] table = new double[nbElements];
          int i = 0;
-         for (float f : information) {
+         for (double f : information) {
             table[i] = f;
             i++;
          }
